Fix up Heia Due
The new stations in Heia Due caused sprite overlap. Moved the first planet closer to the system center so that it doesn't clip with the second station and removed Coppermist because it was so close to the system center that it would clip into the stars. Opted to remove instead of move elsewhere because it was a mothballed station, and we seem to have plenty of those anyway.