Overlap avoidance in Wah Oh + environmental storytelling
Changed the period of Ochrescoop to match the period of the object it orbits, causing the station to always be on the opposite side of the planet relative to the planet's orientation to the star. Wah Oh is an extremely hot star with a high habitability range, so having a station that close to the star could cause problems. Having it be permanently in the shadow of the planet it orbits would help alleviate any such issues.

The periods being as close as they were already is what gave me this idea.