Remove unnecessary comments and change mission names
You should check the wiki for how to make a mission, not have a "template" commented in the data files.

Removed "header" comments for mission regions. Comment headers seem extraneous to me.

Removed "ZZCU" prefix from missions; especially with 1% offer rates, we do NOT need to worry about these missions offering in place of "more important" missions, even if we expect to have way more of these missions in the future.